[LibOS] Implement a checkpoint rwlock to protect access to 'resources'
Implement a checkpoint rwlock that protects the access to resources, such as
PAL_HANDLEs and mmmap'ed memory regions, while a process  performs a checkpoint
due to fork(). While the checkpoint is ongoing other threads must not touch
the resources.

Lock the read lock when
- starting to emulate a syscall
- waking up a thread to die due to exit_group()

Unlock the read lock when
- leaving the emulation of a syscall
- upon execve()

For debugging purposes allow enablement of an atomic counter that is increased
when a read lock is taken and decreased when the read lock is released.

Signed-off-by: Stefan Berger <stefanb@linux.ibm.com>