[python] Add `gramine-sgx-sign --date` option
We did set date in SIGSTRUCT according to system clock without option to
override. When doing reproducible builds, this field needs to be fixed
to a known value, which now can be done using the `--date` option. If no
`--date` option is given, keep the previous behaviour and set the
current date.

SIGSTRUCT keeps dates using BCD, but to avoid messing with setters in
Sigstruct class, this commit won't support setting truly arbitrary
values into the DATE field. Instead, we support only the values that
have only decimal digits in hexadecimal representation.

The tool gramine-sgx-sign is fixed to always output 4 digits for year,
to be copy-pasteable to --date option.

Signed-off-by: Wojtek Porczyk <woju@invisiblethingslab.com>