[LibOS] Introduce 2 states for a thread
Extend the libos_thread with a state field that indicates whether a thread
is currently in a syscall and whether the thread is currently migrating the
state as part of a checkpoint operation. The former can be used to
determine whether to release the checkpoint rea lock before entering a host
syscall where the thread may sleep and the latter can be used to not touch
the checkpoint read lock at all.

Signed-off-by: Stefan Berger <stefanb@linux.ibm.com>