[LibOS] Lock mmap'ed memory area from mmap during checkpointing
When mmap() is run concurrently to a fork() & checkpoint, mmap may add
not-yet-mapped memory areas to the memory mapped areas of a process.
If the checkpoint code wants to transfer them to the destination then
this leads to memory access errors. Therefore, make the addition of
the memory area and the actual memory mapping atomic relative to the
checkpointing.

In particular, make code sequences like the following one atomic
relative to checkpointing:

  bkeep_mmap_xyz()

  ...

  PalVirtualMemoryAlloc()

Signed-off-by: Stefan Berger <stefanb@linux.ibm.com>