[PAL] Refactor splitting of typed URI into type, URI and ops
Previously PAL code had a function `parse_stream_uri()` that returned
the type, URI and corresponding handle operations based on the typed URI
input. This function had confusing signature and semantics. This commit
renames it to `split_uri_and_find_ops()` and makes it explicit that the
input URI is a type-prefixed URI.

Signed-off-by: Dmitrii Kuvaiskii <dmitrii.kuvaiskii@intel.com>