[LibOS] Use static buffer for first thread executing `sendfile()`
This performance optimization uses a statically allocated buffer instead
of allocating on the heap in the `sendfile()` syscall. To prevent data
races on a single static buffer, this optimization is applied only to
the first thread. This optimization is useful for sendfile-heavy
workloads, e.g. Nginx in HTTP mode (and with config `sendfile=on`).

Signed-off-by: Dmitrii Kuvaiskii <dmitrii.kuvaiskii@intel.com>