[LibOS,PAL] Support `EINPROGRESS` on non-blocking sockets connect
Previously, Gramine transformed `connect()` of non-blocking sockets into
a blocking operation and thus never returned -EINPROGRESS. This led to
the connect operation being very slow (waiting for a host timeout) if a
remote peer is unresponsive.

This commit fixes this and adds a LibOS regression test; unfortunately
the test must be run manually with special network settings to properly
test the EINPROGRESS behavior.

Signed-off-by: Dmitrii Kuvaiskii <dmitrii.kuvaiskii@intel.com>