[CI-Examples] Add the BSD-3-clause license to all examples
The CI examples had the default license of LGPL-3.0. With this commit,
we change the license of all Makefiles, manifest templates, config
files, Python scripts in examples to a BSD-3-clause license.

This change is permitted by the "Examples Exception" to the Technical
Charter, enacted on 5. December 2023. This change was agreed by all
contributors to CI examples, past and present.

This change is in response to a popular request to adopt a more
permissive license for our examples. The new BSD-3-clause is a
permissive license (in contrast to the prevoiusly applied copyleft LGPL
license). BSD explicitly allows to modify and use our examples without
any obligation to share these modifications.

Signed-off-by: Dmitrii Kuvaiskii <dmitrii.kuvaiskii@intel.com>