[tools] Disallow nested Gramine invocation
In certain cases, and with bugs in the Gramine build & deployment
scripts, we may end up in Gramine invoking itself (via bash). Since
inner Gramine executable prints hard-to-diagnose error messages in this
case, this commit proactively checks if Gramine executes in the Gramine
environment (by checking for `/dev/attestation/keys/` dir existence),
and prints a human-readable error if it detects such env.

Signed-off-by: Dmitrii Kuvaiskii <dmitrii.kuvaiskii@intel.com>