[LibOS] Set libos_handle free callback while async thread monitors it
Have the async thread set the libos_handle's free callback while it
has the libos_handle on its list, and set the free callback to NULL
once the async thread takes the libos_handle off the list. The callback
allows to remove a libos_handle from the async thread's list once the
libos_handle is supposed to be freed.

Signed-off-by: Stefan Berger <stefanb@linux.ibm.com>