fixup! [common] Introduce pre-allocacted node free list for encrypted files
!TODO: use below commit msg:

[common] Introduce pre-allocacted node free list for encrypted files

Allocation of the Encrypted FS file nodes (e.g., MHT nodes, data nodes)
was found to be a performance bottleneck for certain workloads (e.g.,
RocksDB compaction).

Therefore, instead of using alloc/free directly on file nodes, this
commit introduces a free list of pre-allocacted nodes reserved
specifically for their usage, as a performance optimization that trades
memory for time. A new manifest option has been added to allow users to
set and adjust the free list limit based on their applications.

Note that the encrypted files related tools (`gramine-sgx-pf-crypt`,
`gramine-sgx-pf-tamper`) are not performance-sensitive so we do not
enable this node free list optimization for them. */

Signed-off-by: Kailun Qin <kailun.qin@intel.com>