[LibOS] Refactor eventfd implementation and corresponding tests
This commit refactors eventfd code, in preparation for a second, secure
(not passthrough-to-host) implementation of eventfd. The corresponding
`eventfd` test is split into two tests: a single-process test and a new
multi-process (forking) test. This is because the new secure
implementation of eventfd will only be applicable to single-process
applications.

Signed-off-by: Dmitrii Kuvaiskii <dmitrii.kuvaiskii@intel.com>