[LibOS] Add secure implementation of eventfd
The previous implementation of eventfd was insecure and disabled by
default (enabled via `sys.insecure__allow_eventfd` manifest option). The
new implementation of eventfd is secure but currently restricted to a
single process. Thus, Gramine fails by default if eventfd is used and
fork is issued. However, sometimes it is possible to use this secure
implementation of eventfd in multi-process apps (if child processes do
not use the parent's eventfd object and vice versa); to support these
cases, we introduce the `sys.experimental__allow_eventfd_fork` manifest
option.

LibOS regression tests are added to test the secure implementation of
eventfd. Memcached and Rust examples also use the secure eventfd now.

Signed-off-by: Dmitrii Kuvaiskii <dmitrii.kuvaiskii@intel.com>