[LibOS] Add support for timerfd system calls
This commit adds support for system calls that create and operate on a
timer that delivers timer expiration notifications via a file
descriptor, specifically: `timerfd_create()`, `timerfd_settime()` and
`timerfd_gettime()`. The timerfd object is associated with a dummy
eventfd created on the host to trigger notifications (e.g., in epoll).
The object is created inside Gramine, with all its operations resolved
entirely inside Gramine (note that the time source in Gramine SGX is
still untrusted).

The emulation is currently implemented at the level of a single process.
All timerfds created in the parent process are marked as invalid in
child processes. In multi-process applications, Gramine does not exit
immediately after fork; it only exits if the application attempts to use
timerfds in the child. Therefore, inter-process timing signals via
timerfds are not allowed.

LibOS regression tests are also added.

Signed-off-by: Kailun Qin <kailun.qin@intel.com>