[LibOS] Re-register async event when a periodic itimer expires
Previously, the async event was registered only once on `setitimer()`
syscall and the event's callback blindly reset the interval to 0 when
an interval timer expired. This broke the functionality of a periodic
itimer which fired once rather than periodically.

This commit fixes the above issue by re-regitering an async event in the
itimer event callback if periodic. A LibOS regression test is added.

Signed-off-by: Kailun Qin <kailun.qin@intel.com>