[common] Add 4KB slab size to slab allocator
An encrypted file is split into equally-sized (4KB) chunks (nodes). The
first node is the file metadata node, and the other nodes are stored as
a variant of Merkle Hash Tree (MHT). In this MHT variant, MHT nodes host
the crypto metadata like the encryption keys and MACs of the child
nodes; data nodes host the actual encrypted data and are linked to MHT
nodes.

Allocation of these 4KB file nodes can be one source of overhead for
certain workloads (e.g., RocksDB compaction). To address this
performance bottleneck, this commit introduces a new slab size of 4KB
and therefore a new free list for 4KB blocks managed by the slab
allocator.

Note that while this change can improve performance for certain memory
allocation patterns, it may increase memory overhead and external
fragmentation if the application rarely makes allocations of 4KB.

Signed-off-by: Kailun Qin <kailun.qin@intel.com>