[LibOS] Allow to skip the setup of dentry's inode on lookup
Previously, LibOS unconditionally set up the inode of the dentry on
first lookup, for example, querying the size of the underlying file and
saving it into inode fields. This setup of dentry's inode is not needed
in cases like file removal (`unlink()`), since the inode will be
immediately removed after it was set up. This was particularly painful
for encrypted files (as they would open and decrypt the underlying file
on inode setup) and tmpfs files (as they would allocate an underlying
memory region for file contents on inode setup). Moreover, encrypted
files were decrypted on lookup, just to get the real file size; this
failed `unlink()` on corrupted or encrypted-with-old-key files.

This commit allows to skip the setup of the inode. Currently this is
only used for the `unlink()` system call. This fixes the problematic
case of encrypted files mentioned above and may improve tmpfs
performance. A LibOS test was added.

Signed-off-by: Dmitrii Kuvaiskii <dmitrii.kuvaiskii@intel.com>