[LibOS] Allow `unlink()` on corrupted encrypted files
Currently, LibOS unconditionally sets up the inode of the dentry on
first lookup, for example, querying the size of the underlying file and
saving it into inode fields. This setup of dentry's inode is not needed
in cases like file removal (`unlink()`), since the inode will be
immediately removed after it was set up.

This pecularity made it impossible to remove (unlink) corrupted
encrypted files by the app itself: already the lookup of the file would
fail, disallowing to proceed to the removal operation. "Corrupted" may
mean an attacker-modified file or a more benign case of
encrypted-with-old-key file.

This commit allows to perform `unlink()` on such "corrupted" encrypted
files (by keeping a "shallow" file object). A LibOS test was added.

Signed-off-by: Dmitrii Kuvaiskii <dmitrii.kuvaiskii@intel.com>