[PAL/Linux-SGX] Fix Trusted Files degenerating to Allowed Files on fork
Previously, Trusted Files feature had the following bug: after fork, the
metadata of currently-opened-in-parent-process TFs (SHA256 hashes for
each chunk of the file) was not available in the child SGX enclave. This
effectively degenerated all currently-opened TFs into Allowed Files, and
thus the child enclave lost integrity guarantees in these TFs. An
attacker could substitute the TF contents on the storage with arbitrary
code/data, and the child enclave would not notice this.

This bug was exposed only in the child process (child SGX enclave) and
only in these scenarios:
- fork without execve: the child continues reading from one of the file
  descriptors pointing to a trusted file opened before `fork()`;
- fork with execve: the child reads from an inherited file descriptor
  pointing to a trusted file.

This commit also adds a test to verify this bug is fixed.

Signed-off-by: Dmitrii Kuvaiskii <dmitrii.kuvaiskii@intel.com>