DONTMERGE [PAL/Linux-SGX] Do not call futex OCALL for `futex()` syscalls
App-level `futex()` syscalls end up in `PalEventSet()` and
`PalEventWait()`. If the enclave decides that the condition is not yet
met, then it performs a futex OCALL to instruct the OS kernel to put
this thread to sleep and schedule something else. This commit replaces
the futex OCALL (and thus the blocking sleep) with a busy-poll inside
the enclave.

This way we get rid of one frequent OCALL in multi-threaded apps.

Note that `futex(timeout)` won't benefit from this trick. Also, there is
a caveat that Gramine has 2 helper threads that will also busy-poll, so
one needs 2 additional CPU cores for the whole thing to work (and not
hang indefinitely).

Signed-off-by: Dmitrii Kuvaiskii <dmitrii.kuvaiskii@intel.com>