[LibOS] Move `init_r_debug()` before `init_fs()`
Many subsystems (in particular, their error handling paths) rely on
debug metadata (in particular, the lock for debug info) to be already
initialized. For example, previously the FS subsystem might segfault or
fail on assert in the error handling path (if e.g. `fs.insecure__keys`
manifest option could not be parsed).

Signed-off-by: Dmitrii Kuvaiskii <dmitrii.kuvaiskii@intel.com>