[LibOS] Add `sys.mock_syscalls = [ ... ]` manifest option
This commit adds the manifest syntax `sys.mock_syscalls = [ ... ]`
to specify system calls that will be mocked when executed in Gramine
(i.e. return a specified value without any other side effects).

This may be particularly important for cases where the overhead of
invoking a system call on the host (e.g. exiting the SGX enclave)
becomes a performance bottleneck, and it is more beneficial to disable
or no-op the syscall in the first place; `sched_yield()` is an example.

Another example may be disabling certain functionalities for security
reasons. For example, one may want to disable `eventfd()` and
`eventfd2()` to forbid creation of eventfd objects.

Yet another example may be mocking syscalls currently not implemented in
Gramine. E.g. it may be enough to mock `vhangup()` to always return 0,
so that the workload proceeds further.

Signed-off-by: Dmitrii Kuvaiskii <dmitrii.kuvaiskii@intel.com>