[LibOS] Add `sys.disallowed_syscalls = [ ... ]` manifest option
This commit adds the manifest syntax `sys.disallowed_syscalls = [ ... ]`
to specify system calls that will be disallowed to be executed in
Gramine (i.e. a denylist). This resembles, though significantly less
flexible, seccomp profiles.

This may be particularly important for cases where the overhead of
invoking a system call on the host (e.g. exiting the SGX enclave)
becomes a performance bottleneck, and it is more beneficial to disallow
the syscall in the first place; `sched_yield()` is a good example.

Another example may be disabling certain functionalities for security
reasons. For example, one may want to disable `eventfd()` and
`eventfd2()` to forbid creation of eventfd objects.

Signed-off-by: Dmitrii Kuvaiskii <dmitrii.kuvaiskii@intel.com>