[PAL/Linux-SGX] Re-init attestation key after AESM restarted
Gramine uses the AESM daemon to request SGX Quotes from the Quoting
Enclave. AESM runs in the background and may be restarted. For some
reason, after a restart AESM "forgets" the platform's attestation key
and requires an additional INIT_QUOTE_REQUEST before it is able to
generate SGX Quotes again. This behavior is documented in Intel SGX
PSW/DCAP software, but Gramine didn't implement this and thus failed
with "AESM service returned error 42".

This commit complies with the SGX PSW/DCAP documentation and adds the
send of INIT_QUOTE_REQUEST + retry of SGX quote retrieval.

Signed-off-by: Dmitrii Kuvaiskii <dmitrii.kuvaiskii@intel.com>