[LibOS] Add `stat()` and `hstat()` callbacks to FIFOs
Previously, FIFO pseudo-files (named pipes) did not have callbacks of
`stat()` (used on FIFO filenames) and `hstat()` (used on FIFO
handles/FDs). This breaks some apps that want to double-check that the
file they have opened or want to open is a FIFO, using the `stat()` or
`fstat()` syscalls. There was no good reason to skip those callbacks, so
this commit adds their dummy implementations.

Signed-off-by: Dmitrii Kuvaiskii <dmitrii.kuvaiskii@intel.com>