fixup! [LibOS,PAL] Emulate file-backed mmap via PAL read/write APIs
Per suggestion from @mwk, this fixup adds `vma->valid_length` field. In
this way, we commit to the model of "logical" split within a single VMA:
some prefix of the VMA is actually active (and has valid pages), while
the rest is unmapped (and returns SIGBUS). Only file-backed VMAs are
split in this way (anonymous-memory VMAs can't be in "unmapped" state).

Signed-off-by: Dmitrii Kuvaiskii <dmitrii.kuvaiskii@intel.com>