[tools/RA-TLS] Copy quote from X.509 cert into a separate object
Previously, `extract_quote_and_verify_pubkey()` returned a pointer to
the SGX quote located inside the X.509 certificate. This is a confusing
pattern, so this commit introduces a copy operation, to copy the SGX
quote into a newly allocated object whose ownership is passed to the
callers of this func.

Signed-off-by: Dmitrii Kuvaiskii <dmitrii.kuvaiskii@intel.com>