[PAL,LibOS,common] Enforce that PAL_ERROR is always negative
This commit asserts that we always use negative PAL_ERROR_* values,
and as a result also deduplicates the pal_linux_error.h file, which was
identical in both PAL/Linux and PAL/Linux-SGX. The file is now located
in the linux-common directory.

Also, this commit adds an assert on mbedTLS errors being
negative.

Signed-off-by: Mariusz Zaborski <oshogbo@invisiblethingslab.com>