[PAL/Linux-SGX] Error out on enclave_dbginfo allocation failure
We're modifying this structure dynamically since we introduced support
for dynamic TCS. The modification code in add_dynamic_tcs() has no way
of knowing whether the allocation failed or succeeded, thus it has no
way to decide whether writing there is legal or not.

This commit guarantees that the stucture is always allocated.

Signed-off-by: Michał Kowalczyk <mkow@invisiblethingslab.com>