[LibOS] Fix bug of `RLIMIT_STACK` being overwritten in child
Previously, the resource limit `RLIMIT_STACK` (maximum size of the
process stack, in bytes) was overwritten in the child process, after
fork. Thus, if the parent used `setrlimit(RLIMIT_STACK)`, the effect of
this was lost in the child. This commit fixes this bug, and adds a LibOS
test to verify the correct behavior.

Signed-off-by: Dmitrii Kuvaiskii <dmitrii.kuvaiskii@intel.com>