[LibOS] Make async worker persistent
The async worker thread is currently spawned on-demand and terminated
after an idle timeout. This is entirely unnecessary — the cost of having
a persistent async thread is near-0, and it simplifies the logic.

Also fixes a long-standing issue about not waiting for the async thread
to exit (https://github.com/gramineproject/graphene/issues/440).