[PAL] Fix `memory` PAL regression test for UBSan
Newer Clang versions (v18 and newer) added more UBSan checks that
triggered UBSan on PAL regression tests. In particular, UBSan added the
check "Indirect call of a function through a function pointer of the
wrong type". UBSan expects all functions which can be indirectly called
to be instrumented with two magic metadata values, located right-before
the function in address space.

The `memory` test however does *not* add this required metadata to a
dummy generated-on-the-fly function (it simply allocates some pages and
copies the asm code at the beginning of the first page). Instead of
adding UBSan-required metadata, we simply disable this particular check
(it's a test after all, not core Gramine functionality).

Signed-off-by: Dmitrii Kuvaiskii <dmitrii.kuvaiskii@intel.com>