[PAL] Fix `attestation` LibOS regression test for UBSan
UBSan detects an uninitialized boolean variable on a DCAP SGX machine on
the remote attestation example, failing with the message:

    error: ubsan: load of invalid value for bool or enum: 100
    error: ubsan: ../pal/src/host/linux-sgx/pal_misc.c:724:43

This happens because `bool linkable` variable is assigned only in the
EPID attestation scheme, but left unassigned in the DCAP scheme. This
commit fixes this (though it's not a bug since DCAP never uses
`linkable` anyway). Hence, UBSan doesn't complain anymore.

Signed-off-by: Dmitrii Kuvaiskii <dmitrii.kuvaiskii@intel.com>